% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_params.R
\docType{data UPDATE!!!}
\name{example_params}
\alias{example_params}
\title{A sample SingleCellExperiment dataset generated from a dose-response
experiments used to illustrate dose-response data simulation.}
\format{
a matrix instance, 1 row per gene. UPDATE!!!
}
\source{
Wellcome Trust Centre for Human Genetics, Oxford UPDATE!!!
}
\usage{
example_sc_cholangiocytes UPDATE!!!
}
\value{
NULL, but makes aavailable a matrix of count data UPDATE!!!
}
\description{
This data set contains counts for 2000 genes for 40 cells. They UPDATE!!!
are from a real experiment, but details have been anonymised. UPDATE!!!
}
\author{
Davis McCarthy, 2015-03-05 UPDATE!!!
}
