% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response_models.R
\name{modelExp}
\alias{modelExp}
\title{Generate count values following a Exponential (2 or 3) model
response = a * exp(sign * (b * dose)^d)}
\usage{
modelExp(doses, a, b, d)
}
\arguments{
\item{doses}{A vector of doses to model}

\item{a}{The background response}

\item{b}{'Slope' of the model}

\item{d}{'Power' of the model. Set as 1 for Exponential 2}
}
\description{
Generate count values following a Exponential (2 or 3) model
response = a * exp(sign * (b * dose)^d)
}
