% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response_models.R
\name{modelExpB}
\alias{modelExpB}
\title{Generate count values following a Exponential (2 or 3) model
response = a(c-(c-1) * exp(-1 (b * dose)^d))}
\usage{
modelExpB(doses, a, b, c, d)
}
\arguments{
\item{doses}{A vector of doses to model}

\item{a}{The background response}

\item{b}{'Slope' of the model}

\item{c}{Asymptote where 0 < c < 1 is decreasing data. c > 0.}

\item{d}{'Power' of the model}
}
\description{
Generate count values following a Exponential (2 or 3) model
response = a(c-(c-1) * exp(-1 (b * dose)^d))
}
