% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{data}
\name{params}
\alias{params}
\title{A sample Splatter parameter object generated from a real dose-response
single-nuclei RNA sequencing experiment.}
\format{
SplatParams
}
\source{
Michigan State University, East Lansing, MI
}
\usage{
data(params)
}
\value{
SplatParams object
}
\description{
This parameter object was generated using the \code{splatEstimate}
function of \code{Splatter} on real dose-response single-nuclei RNA sequencing
data. The number of genes was changed to 2000 and number of cells to 4500 to
serve as template for generating dose-response simulated data consisting of
9 dose groups. The number of batches was updated to 27 to represent 3
biological replicates for each dose group.
}
\author{
Rance Nault, 2021-08-06
}
\keyword{datasets}
