% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plotGeneQuantiles}
\alias{plotGeneQuantiles}
\title{Generate count values following a Hill model
response = ?? + (V * dose^n)/(k^n + dose^n)}
\usage{
plotGeneQuantiles(sce, feature)
}
\arguments{
\item{doses}{A vector of doses to model}

\item{gamma}{The background response}

\item{V}{Velocity, or the maximal response}

\item{n}{Power of the model. Should be > 1 to avoid an infinite slope.}

\item{k}{Dose were half the response is observed. Equivale of ED/EC50}
}
\description{
Generate count values following a Hill model
response = ?? + (V * dose^n)/(k^n + dose^n)
}
