% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{RMSE}
\alias{RMSE}
\title{Calculate root mean square error statistics}
\usage{
RMSE(m, o)
}
\arguments{
\item{m}{A numeric vector to calculate distance from}

\item{o}{A numeric vector to calculate the distance to}
}
\description{
Calculate root mean square error statistics
}
