% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions_splatter.R
\name{simUnchanged}
\alias{simUnchanged}
\title{Simulate expression that follows an power model
response = gamma + beta * doses^delta}
\usage{
simUnchanged(doses, realVec)
}
\arguments{
\item{doses}{A vector of doses to model}

\item{realVec}{A vector of means fear each dose group of equal length as dose}
}
\value{
a list of the model fit parameters including a resp vector containing
values for each dose group.
}
\description{
Simulate expression that follows an power model
response = gamma + beta * doses^delta
}
\examples{
simUnchanged(c(0, 1, 3, 10, 30), rep(1, 5))
}
