% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions_splatter.R
\name{splatsimExpB}
\alias{splatsimExpB}
\title{Simulate expression that follows an exponential model (2 or 3)
response = a(c-(c-1)exp???(-1 (bdose)^d ))}
\usage{
splatsimExpB(doses, mean, fc = 1.5, verbose = FALSE)
}
\arguments{
\item{doses}{A vector of doses to model}

\item{mean.range}{A vector of means obtain from realData}

\item{fc.range}{a vector with the minimum and maximim |fold-change| (e.g., c(1.5, 5))}

\item{downregulated}{set a TRUE to model repression instead of induction}
}
\value{
a list of the model fit parameters including
}
\description{
Simulate expression that follows an exponential model (2 or 3)
response = a(c-(c-1)exp???(-1 (bdose)^d ))
}
\examples{
simExpB(c(0,1,3,10,30), realData$mean)
}
