% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions_splatter.R
\name{splatsimLinear}
\alias{splatsimLinear}
\title{Simulate expression that follows a linear}
\usage{
splatsimLinear(doses, mean, fc, ...)
}
\arguments{
\item{doses}{A vector of doses to model}

\item{mean}{A number representing the value at dose 0}

\item{fc}{A number representing the maximum fold-change}

\item{...}{Additional arguments}
}
\value{
a list of the model fit parameters including a resp vector containing
values for each dose group and the fit parameters
}
\description{
Simulate expression that follows a linear
}
\examples{
splatsimLinear(c(0, 1, 3, 10, 30), mean = 1, fc = 1.5)
}
