% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splatterWrapper.R
\name{splatSimulateDR}
\alias{splatSimulateDR}
\title{Simulate dose-response data}
\usage{
splatSimulateDR(
  params = newSplatParams(),
  method = c("doseresp"),
  sparsify = TRUE,
  verbose = TRUE,
  dose.names = c(0),
  dose.prob = c(1),
  lib.scale = 1.3,
  ...
)
}
\arguments{
\item{method}{}
}
\value{
sce
}
\description{
Add some details here
}
\examples{

}
\seealso{

}
