% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{summarizeGene}
\alias{summarizeGene}
\title{Calculate gene-wise summary values from \code{SingleCellExperiment} object}
\usage{
summarizeGene(sce, dataset.type, dataset.name)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}

\item{dataset.type}{A character naming the dataset type (e.g., 'real data')}

\item{dataset.name}{A character naming the dataset (e.g., 'dose-response')}
}
\description{
Calculate gene-wise summary values from \code{SingleCellExperiment} object
}
