% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response_models.R
\name{modelPolynomial}
\alias{modelPolynomial}
\title{Generate count values following a polynomial model
response = beta_0 + beta_(1) * dose + beta_(2) * dose^2+...+ beta(n) * dose^n}
\usage{
modelPolynomial(doses, gamma, beta)
}
\arguments{
\item{doses}{A vector of doses to model}

\item{gamma}{The background response}

\item{beta}{A vector of 'Slopes' for the model. Length of 1 means linear.}
}
\description{
Generate count values following a polynomial model
response = beta_0 + beta_(1) * dose + beta_(2) * dose^2+...+ beta(n) * dose^n
}
