% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions_splatter.R
\name{simUnchanged}
\alias{simUnchanged}
\title{Simulate expression that follows an power model
response = gamma + beta * doses^delta}
\usage{
simUnchanged(doses, realVec)
}
\arguments{
\item{doses}{A vector of doses to model}

\item{mean.range}{A vector of means obtain from realData}

\item{fc.range}{a vector with the minimum and maximim |fold-change| (e.g., c(1.5, 5))}

\item{downregulated}{set a TRUE to model repression instead of induction}
}
\value{
a list of the model fit parameters including
}
\description{
Simulate expression that follows an power model
response = gamma + beta * doses^delta
}
\examples{
simPower(c(0,1,3,10,30), realData$mean)
}
